+++ src/main/java/com/jin/env/garbage/entity/point/GarbageUserPointEntity.java
@@ -18,12 +18,12 @@ public class GarbageUserPointEntity  extends BaseEntity{
     private Integer point;
     private String phone;
 
-    private Integer provinceId;
+    private Long provinceId;
     private Long cityId;
     private Long countryId;
     private Long townId;
     private Long villageId;
-    private Integer communityId;
+    private Long communityId;
     private String communityName;
 
     @Basic
@@ -129,11 +129,11 @@ public class GarbageUserPointEntity  extends BaseEntity{
 
     @Basic
     @Column(name = "province_id")
-    public Integer getProvinceId() {
+    public Long getProvinceId() {
         return provinceId;
     }
 
-    public void setProvinceId(Integer provinceId) {
+    public void setProvinceId(Long provinceId) {
         this.provinceId = provinceId;
     }
 
@@ -179,11 +179,11 @@ public class GarbageUserPointEntity  extends BaseEntity{
 
     @Basic
     @Column(name = "community_id")
-    public Integer getCommunityId() {
+    public Long getCommunityId() {
         return communityId;
     }
 
-    public void setCommunityId(Integer communityId) {
+    public void setCommunityId(Long communityId) {
         this.communityId = communityId;
     }
